+++ src/main/java/kroppeb/stareval/parser/Parser.java
@@ -1,127 +1,99 @@
 package kroppeb.stareval.parser;
 
-
-import kroppeb.stareval.exception.*;
+import kroppeb.stareval.exception.MissingTokenException;
+import kroppeb.stareval.exception.ParseException;
+import kroppeb.stareval.exception.UnexpectedTokenException;
 import kroppeb.stareval.token.*;
-import net.minecraft.client.util.CharPredicate;
-import org.jetbrains.annotations.NotNull;
 
+import java.util.ArrayList;
+import java.util.List;
 
-public class Parser {
-	private final ParserOptions options;
-	private final ParserOptions.ParserParts parserParts;
+class Parser {
+	final List<Token> stack = new ArrayList<>();
 
-	public Parser(ParserOptions options) {
-		this.options = options;
-		this.parserParts = options.getParserParts();
-	}
-
-	public ExpressionToken parse(String input) throws ParseException {
-		return this.parseInternal(new StringReader(input));
-	}
-
-	private ExpressionToken parseInternal(StringReader input) throws ParseException {
-		// parser stack
-		final TokenStack stack = new TokenStack();
-
-
-		while (input.canRead()) {
-			char c = input.read();
-
-			if (this.parserParts.isIdStart(c)) {
-				Token token = this.parseIdGroup(input);
-				stack.push(token);
-			} else if (this.parserParts.isNumberStart(c)) {
-				// start parsing a number
-				final String numberString = readWhile(input, this.parserParts::isNumberPart);
-				stack.push(new NumberToken(numberString));
-			} else if (c == '(') {
-				stack.push(new UnfinishedArgsToken());
-			} else if (c == ',') {
-				if (stack.peek() instanceof ExpressionToken) {
-					stack.commaReduce(input.getCurrentIndex());
-				} else {
-					throw new UnexpectedTokenException("Expected an expression before a comma ','", input.getCurrentIndex());
-				}
-			} else if (c == ')') {
-				stack.bracketReduce(input.getCurrentIndex());
-			} else {
-				if (stack.peek() instanceof ExpressionToken) {
-					// maybe binary operator
-					OpResolver<? extends BinaryOp> resolver = this.options.getBinaryOpResolver(c);
-
-					if (resolver != null) {
-						stack.push(new BinaryOperatorToken(resolver.resolve(input)));
-						continue;
-					}
-				} else {
-					// maybe unary operator
-					OpResolver<? extends UnaryOp> resolver = this.options.getUnaryOpResolver(c);
-
-					if (resolver != null) {
-						stack.push(new UnaryOperatorToken(resolver.resolve(input)));
-						continue;
-					}
-				}
-
-				throw new UnexpectedCharacterException(c, input.getCurrentIndex());
-			}
+	Token peek() {
+		if (!this.stack.isEmpty()) {
+			return this.stack.get(this.stack.size() - 1);
 		}
 
-		if (!stack.stack.isEmpty()) {
-			if (stack.peek() instanceof ExpressionToken) {
-				ExpressionToken result = stack.expressionReducePop();
+		return null;
+	}
 
-				if (stack.stack.isEmpty()) {
-					return result;
-				}
+	private Token pop() {
+		if (this.stack.isEmpty()) {
+			throw new IllegalStateException("Internal token stack is empty");
+		}
 
-				if (stack.peek() instanceof UnfinishedArgsToken) {
-					throw new MissingTokenException("Expected a closing bracket", input.getCurrentIndex());
-				} else {
-					throw new UnexpectedTokenException(
-							"The stack of tokens isn't empty at the end of the expression: " + stack.stack +
-									" top: " + result, input.getCurrentIndex());
-				}
-			} else {
-				Token top = stack.peek();
-				if (top instanceof UnfinishedArgsToken) {
-					throw new MissingTokenException("Expected a closing bracket", input.getCurrentIndex());
-				} else if (top instanceof PriorityOperatorToken) {
-					throw new MissingTokenException(
-							"Expected a identifier, constant or subexpression on the right side of the operator",
-							input.getCurrentIndex());
-				} else {
-					throw new UnexpectedTokenException(
-							"The stack of tokens contains an unexpected token at the top: " + stack.stack,
-							input.getCurrentIndex());
-				}
-			}
+		return this.stack.remove(this.stack.size() - 1);
+	}
+
+	/**
+	 * Executes following reduce steps:
+	 * <ul>
+	 *     <li>{@link ExpressionToken} | {@link BinaryOperatorToken} => {@link PartialBinaryExpressionToken}</li>
+	 *     <li>{@link UnaryOperatorToken}, {@link ExpressionToken} | {@link BinaryOperatorToken} => {@link UnaryExpressionToken} | {@link BinaryOperatorToken}</li>
+	 *     <li>
+	 *         {@link PartialBinaryExpressionToken}, {@link ExpressionToken} | {@link BinaryOperatorToken} <br/>
+	 *         where the operator on the stack has a higher or equal priority to the one being added, the 3 items on the
+	 *         stack get popped, merged to a {@link BinaryExpressionToken} and placed on the stack.
+	 *         The new token is then pushed again.
+	 *     </li>
+	 * </ul>
+	 */
+	void push(Token token) {
+		final Token top = this.peek();
+
+		if (token instanceof BinaryOperatorToken) {
+			BinaryOperatorToken binOpToken = (BinaryOperatorToken) token;
+
+			// reduce the expressions to the needed priority level
+			ExpressionToken left = this.expressionReducePop(binOpToken.op.getPriority());
+			// stack[ {'a', '*'}, 'b'], token = '+' -> stack[], left = {'a', '*', 'b'}
+			//                                      -> stack[{{'a', '*', 'b'}, '+'}]
+			// stack[ {'a', '+'}, 'b'], token = '+' -> stack[], left = {'a', '+', 'b'}
+			//                                      -> stack[{{'a', '+', 'b'}, '+'}]
+			// stack[ {     '-'}, 'b'], token = '+' -> stack[], left = {'-', 'b'}
+			//                                      -> stack[{{     '-', 'b'}, '+'}]
+
+			// stack[ {'a', '+'}, 'b'], token = '*' -> stack[{'a', '+'}], left = {'b'}
+			//                                      -> stack[{'a', '+'}, {'b', '*'}]
+
+			this.stack.add(new PartialBinaryExpressionToken(left, binOpToken.op));
+
 		} else {
-			throw new MissingTokenException("The input seems to be empty", input.getCurrentIndex());
+			this.stack.add(token);
 		}
 	}
 
-	@NotNull
-	private Token parseIdGroup(StringReader input) throws ParseException {
-		final String id = readWhile(input, this.parserParts::isIdPart);
-		AccessableToken token = new IdToken(id);
+	/**
+	 * @throws ClassCastException if the top is not an expression
+	 * @see #expressionReducePop(int)
+	 */
+	ExpressionToken expressionReducePop() {
+		return this.expressionReducePop(Integer.MAX_VALUE);
+	}
 
-		if (!input.canRead()) {
-			return token;
-		}
+	/**
+	 * Pops an expression after trying to reduce the stack.
+	 * Executes following reduce steps:
+	 * <ul>
+	 *     <li>{@link PriorityOperatorToken}, {@link ExpressionToken} => {@link ExpressionToken}
+	 *     as long as the {@code priority} of the {@link PriorityOperatorToken} is stricter than the given priority</li>
+	 * </ul>
+	 *
+	 * @throws ClassCastException if the top is not an expression
+	 */
+	private ExpressionToken expressionReducePop(int priority) {
+		ExpressionToken token = (ExpressionToken) this.pop();
 
+		while (!this.stack.isEmpty()) {
+			Token x = this.peek();
 
-		while (input.tryRead('.')) {
-			if (input.canRead()) {
-				char start = input.read();
-				if (!this.parserParts.isAccessStart(start)) {
-					throw new UnexpectedCharacterException("a valid accessor", start, input.getCurrentIndex());
-				}
-
-				token = new AccessToken(token, readWhile(input, this.parserParts::isAccessPart));
+			if (x instanceof PriorityOperatorToken && ((PriorityOperatorToken) x).getPriority() <= priority) {
+				this.pop();
+				token = ((PriorityOperatorToken) x).resolveWith(token);
 			} else {
-				throw new UnexpectedEndingException("An expression can't end with '.'");
+				break;
 			}
 		}
 
@@ -129,21 +101,91 @@ public class Parser {
 	}
 
 	/**
-	 * The returned value includes the last value btw;
+	 * Executes following reduce step:
+	 * <ul>
+	 *     <li>{@link UnfinishedArgsToken}, {@link #expressionReducePop} => {@link UnfinishedArgsToken}</li>
+	 * </ul>
+	 *
+	 * @param index the current reader index, for exception throwing.
+	 *
+	 * @throws ClassCastException if the top is not an expression
 	 */
-	private static String readWhile(StringReader input, CharPredicate predicate) {
-		input.mark();
+	void commaReduce(int index) throws ParseException {
+		// ( expr,
+		// UnfinishedArgs Expression (commaReduce)
+		// => UnfinishedArgs
 
-		while (input.canRead()) {
-			if (!predicate.test(input.peek())) {
-				break;
-			}
+		ExpressionToken expr = this.expressionReducePop();
+		Token args = this.peek();
 
-			input.skip();
+		if (args == null) {
+			throw new MissingTokenException(
+					"Expected an opening bracket '(' before seeing a comma ',' or closing bracket ')'",
+					index
+			);
 		}
 
-		return input.substring();
+		if (args instanceof UnfinishedArgsToken) {
+			((UnfinishedArgsToken) args).tokens.add(expr);
+		} else {
+			throw new UnexpectedTokenException(
+					"Expected to see an opening bracket '(' or a comma ',' right before an expression followed by a " +
+							"closing bracket ')' or a comma ','", index);
+		}
+	}
+
+	/**
+	 * Allows for trailing comma.
+	 * Executes following reduce steps:
+	 * <ul>
+	 *     <li>{@link IdToken} {@link UnfinishedArgsToken}, {@link #expressionReducePop} => {@link CallToken}</li>
+	 *     <li>{@link IdToken} {@link UnfinishedArgsToken} => {@link CallToken}</li>
+	 *     <li>{@link UnfinishedArgsToken}, {@link #expressionReducePop} => {@link ExpressionToken}</li>
+	 * </ul>
+	 *
+	 * @param index the current reader index, for exception throwing.
+	 */
+	void bracketReduce(int index) throws ParseException {
+		//
+		// ( ... )
+		// UnfinishedArgsToken Expression? (callReduce)
+
+		boolean expressionOnTop = this.peek() instanceof ExpressionToken;
+		if (expressionOnTop) {
+			this.commaReduce(index);
+		}
+
+		UnfinishedArgsToken args;
+		{
+			if(this.stack.isEmpty()){
+				throw new MissingTokenException("A closing bracket ')' can't be the first character of an expression", index);
+			}
+
+			Token pop = this.pop();
+
+			if (!(pop instanceof UnfinishedArgsToken)) {
+				throw new UnexpectedTokenException(
+						"Expected to see an opening bracket '(' or a comma ',' right before an expression followed by a " +
+								"closing bracket ')' or a comma ','", index);
+			}
+			args = (UnfinishedArgsToken) pop;
+		}
+
+		Token top = this.peek();
+
+		if (top instanceof IdToken) {
+			this.pop();
+			this.push(new CallToken(((IdToken) top).id, args.tokens));
+		} else {
+			if(args.tokens.isEmpty()){
+				throw new MissingTokenException("Encountered empty brackets that aren't a call", index);
+			} else if(args.tokens.size() > 1){
+				throw new UnexpectedTokenException("Encountered too many expressions in brackets that aren't a call", index);
+			} else if(!expressionOnTop){
+				throw new UnexpectedTokenException("Encountered a trailing comma in brackets that aren't a call", index);
+			} else {
+				this.push(args.tokens.get(0));
+			}
+		}
 	}
 }
-
-
